Refactor: JwtUtil 추상 클래스로 수정
