+++ ilkda/src/main/java/com/ilkda/server/auth/service/AuthService.java
@@ -92,7 +92,7 @@ public class AuthService {
             throw new UnauthorizedException("ACCESS 토큰을 갱신할 수 없습니다.");
         }
 
-        Long memberId = memberJwtUtil.getPayload().getMember_id();
+        Long memberId = ((MemberJwtPayload)memberJwtUtil.getPayload()).getMember_id();
         if(!memberService.existsMember(memberId)) {
             log.debug("{} 멤버 존재 x", memberId);
             throw new NotFoundException("존재하지 않는 회원입니다.");
@@ -114,7 +114,7 @@ public class AuthService {
                     MemberJwtPayload.builder()
                     .type(JwtPayload.JwtType.REFRESH)
                     .exp(System.currentTimeMillis() + REFRESH_TOKEN_DURATION)
-                    .member_id(memberJwtUtil.getPayload().getMember_id())
+                    .member_id(((MemberJwtPayload)memberJwtUtil.getPayload()).getMember_id())
                     .build());
         }
         return refreshToken;
