+++ ilkda/src/main/java/com/ilkda/server/jwt/util/MemberJwtUtil.java
@@ -9,31 +9,17 @@ import org.json.simple.parser.ParseException;
 
 import java.util.*;
 
-public class MemberJwtUtil implements JwtUtil {
-
-    private String token;
-    private MemberJwtPayload payload;
+public class MemberJwtUtil extends JwtUtil {
 
     public MemberJwtUtil(String token) {
-        this.token = token;
-        this.payload = parsePayload();
-
-        validateExpiration(payload.getExp());
-    }
-
-    @Override
-    public String getToken() {
-        return this.token;
-    }
-
-    public MemberJwtPayload getPayload() {
-        return this.payload;
+        super(token);
     }
 
     /**
      * jwt의 {헤더.페이로드.서명} 형식에서 페이로드를 가져오기 위해 [.]을 기준으로 split 후 두 번째 인자를 사용합니다. <br/>
      * jwt는 HS256 (Base64(Header) + Base64(Payload) + secret key) 알고리즘으로 인코딩 되었으므로 Base64로 페이로드를 디코딩 합니다.
      * */
+    @Override
      public MemberJwtPayload parsePayload() {
         try {
             JSONParser jsonParser = new JSONParser();
